requirements.txt
transformers
accelerate
qwen_vl_utils
pillow
numpy
huggingface_hub